Struct linera_alloy_contract::ContractInstance

pub struct ContractInstance<T, P, N = Ethereum> { /* private fields */ }

A handle to an Ethereum contract at a specific address.

A contract is an abstraction of an executable program on Ethereum. Every deployed contract has an address, which is used to connect to it so that it may receive messages (transactions).

Implementations

impl<T, P, N> ContractInstance<T, P, N>

pub const fn new(address: Address, provider: P, interface: Interface) -> Self

Creates a new contract from the provided address, provider, and interface.

pub const fn address(&self) -> &Address

Returns a reference to the contract’s address.

pub fn set_address(&mut self, address: Address)

Sets the contract’s address.

pub fn at(self, address: Address) -> Self

Returns a new contract instance at address.

pub const fn abi(&self) -> &JsonAbi

Returns a reference to the contract’s ABI.

pub const fn provider(&self) -> &P

Returns a reference to the contract’s provider.

impl<T, P: Clone, N> ContractInstance<T, &P, N>

pub fn with_cloned_provider(self) -> ContractInstance<T, P, N>

Clones the provider and returns a new contract instance with the cloned provider.

impl<T: Transport + Clone, P: Provider<T, N>, N: Network> ContractInstance<T, P, N>

pub fn function( &self, name: &str, args: &[DynSolValue], ) -> Result<CallBuilder<T, &P, Function, N>>

Returns a transaction builder for the provided function name.

If there are multiple functions with the same name due to overloading, consider using the ContractInstance::function_from_selector method instead, since this will use the first match.

pub fn function_from_selector( &self, selector: &Selector, args: &[DynSolValue], ) -> Result<CallBuilder<T, &P, Function, N>>

Returns a transaction builder for the provided function selector.

pub const fn event<E: SolEvent>(&self, filter: Filter) -> Event<T, &P, E, N>

Returns an Event builder with the provided filter.